fix(audit): PR #400 review + audit findings
Direction-bound bindings: 11 `Behavior on X { profile: f(X) }` sites
re-evaluated mid-animation as the property interpolated, picking the
wrong leg at start. Replaced with intent-driven conditionals reading
non-animated state (visible / fullscreenMode / sidebarCompact /
needsSave / etc.) across EditorWindow, EditorZone, PropertyPanel,
Main, LayoutCard, ZonePreview.

Dead paths removed: `popup.snapAssist.hide` (surface destroys-on-hide,
no frame paints) and `zone.layoutSwitchOut` (no consumer). Kept as
"intentionally absent — restore in lockstep with consumer" comments.

Test fixture realigned: test_overlay_scope_prefixes showScaleProfile
strings now match production (`osd.show` and `popup.layoutPicker.show`,
not the divergent `osd.pop` / `popup.layoutPicker.popIn`).

Comment regressions: `pop` vs `popup` substring example in
isAncestorOrSelf and the descendant-prefix comment in
animationspagecontroller now correctly illustrate the dot-terminator
bug they document.

QML housekeeping: rewrite misleading "DEPRECATED — vestigial" comments
on actively-used animationDuration properties; replace hardcoded
150/200 ms with Kirigami.Units.shortDuration/longDuration; wrap `/2`
durationOverride sites with Math.round.

File splits (project <800-line guideline):
- test_settings_core.cpp 1390 → 688 + test_settings_animation_profile
  (449) + test_settings_shader_tree (338)
- test_animations_page_controller.cpp 1032 → 499 +
  test_animations_motion_sets (451) + test_animations_shader_overrides
  (156)
- overlayservice.cpp 1533 → 1015, with build*Config + animator wireup
  extracted to overlayservice/animation_config.cpp (314) and
  show/hide/toggle/idle to overlayservice/lifecycle.cpp (279),
  alongside existing selector/snapassist/osd siblings.

Stale historical-context comments in selector.cpp,
ZoneSelectorWindow.qml, and snapassist.cpp updated to reference the
actual current path strings.
