+++ src/daemon/overlayservice.cpp
@@ -115,286 +115,10 @@ void cleanupVirtualScreenStates(QHash<QString, OverlayService::PerScreenOverlayS
 
 } // namespace
 
-namespace {
-
-// ── Per-role Config builders ───────────────────────────────────────────
-// One factory per overlay role so adding/tweaking a 6th overlay touches
-// exactly one named function rather than appending to a 100-line
-// setupSurfaceAnimator. Each builder documents the visual shape it
-// encodes; setupSurfaceAnimator just wires the builder output to the
-// matching PzRole.
-//
-// **Surface-family separation policy.** Profile paths are partitioned by
-// surface family so each overlay tunes independently:
-//
-//   - **OSD family (`osd.*`)** — genuine OSDs only. Used by the
-//     Notification surface (LayoutOsd / NavigationOsd). A JSON edit to
-//     `osd.hide` affects OSDs and ONLY OSDs.
-//   - **Popup family (`popup.<surface>.*`)** — non-OSD overlay
-//     surfaces. Each gets its own leaf paths under
-//     `popup.<surface>` per surface's needs:
-//       • `popup.layoutPicker.{show, hide, popIn}` (opacity legs +
-//         scale show; hide-scale couples to `.hide`)
-//       • `popup.zoneSelector.{show, hide}` (opacity-only)
-//       • `popup.snapAssist.show` (destroy-on-hide; no `.hide`
-//         leaf because no frame ever paints)
-//     A JSON edit to `popup.layoutPicker.hide` affects ONLY the
-//     layout picker; siblings are unaffected.
-//
-// The shipped tree carries zero bundled per-leaf profile JSONs — every
-// profile is sourced from the Settings UI's per-node overrides via
-// `PhosphorProfileRegistry::registerProfile`, with
-// `resolveWithInheritance()` walking the parent chain so a parent-node
-// edit (e.g. "All Popups → 2000 ms" written to `popup`)
-// propagates to every leaf under it. Unset paths fall through to
-// library defaults (150 ms OutCubic). User-authored JSONs at
-// `~/.local/share/plasmazones/profiles/<path>.json` are still loaded
-// by ProfileLoader for advanced users who want file-based overrides.
-//
-// **Within-family scale-leg coupling (intentional, scoped).** Each
-// surface family's hide-leg-scale reuses the surface family's
-// hide-leg-opacity path (e.g. `osd.hide` drives both opacity and scale
-// hide for OSDs). Editing `osd.hide` to add Spring physics affects both
-// legs of the OSD hide. This is the same pattern OSDs and (formerly)
-// LayoutPicker shared — kept INSIDE each surface family but not BETWEEN
-// them, which is the change here. If a future surface needs decoupled
-// opacity/scale tuning, introduce a sibling path
-// (e.g. `popup.layoutPicker.popOut`) and register it on
-// `hideScaleProfile`.
-
-namespace PAL = PhosphorAnimationLayer;
-namespace PAS = PhosphorAnimationShaders;
-
-/// Resolve a path against the shader profile tree. A default-constructed
-/// tree (empty baseline + no overrides) resolves every path to an empty
-/// effect id, equivalent to "no shader leg" — motion runs alone, identical
-/// to the pre-shader-wireup behaviour. The setSettings() handler later
-/// re-registers configs with the live tree once settings exist.
-///
-/// **Source-of-truth note.** The settings UI gates its shader picker on
-/// `src/core/animationshadersupportedpaths.h::shaderSupportedEventPaths`,
-/// which enumerates exactly the paths consumed by `resolveShaderEffect`
-/// in the build*Config functions below. When a new shader-leg surface
-/// lands here, append its leg paths to that list in lockstep so the
-/// settings UI starts surfacing the picker on the new path.
-QString resolveShaderEffect(const PAS::ShaderProfileTree& tree, const QString& path)
-{
-    return tree.resolve(path).effectiveEffectId();
-}
-
-/// Resolve a path against the shader profile tree and extract the per-event
-/// parameter overrides. Empty map when the profile didn't override anything
-/// — the shader runs with its declared defaults from metadata.json.
-QVariantMap resolveShaderParameters(const PAS::ShaderProfileTree& tree, const QString& path)
-{
-    return tree.resolve(path).effectiveParameters();
-}
-
-/// Default config — empty. Surfaces that route through the animator
-/// without a registered config fall back to AnimatedValue's library
-/// default (150 ms OutCubic), same as a missing-profile lookup. Every
-/// Surface that goes through Surface::show()/hide() in this service has
-/// an explicit registration below; the empty default is documentation.
-PAL::SurfaceAnimator::Config buildDefaultConfig()
-{
-    return PAL::SurfaceAnimator::Config{};
-}
-
-/// LayoutOsd / NavigationOsd: identical fade-and-pop shape. Shader leg
-/// resolves osd.show / osd.hide from the user's tree so OSDs can fly in
-/// (slide), dissolve, etc. independently of zone or popup events.
-///
-/// **Genuine-OSD surface family.** All four leg paths live under
-/// `osd.*`. A JSON edit to `osd.hide` affects OSDs and ONLY OSDs —
-/// LayoutPicker / ZoneSelector / SnapAssist live in the
-/// `popup.*` family.
-PAL::SurfaceAnimator::Config buildOsdConfig(const PAS::ShaderProfileTree& tree)
-{
-    namespace PP = PhosphorAnimation::ProfilePaths;
-    return PAL::SurfaceAnimator::Config{.showProfile = PP::OsdShow,
-                                        .hideProfile = PP::OsdHide,
-                                        .showScaleProfile = PP::OsdShow,
-                                        .hideScaleProfile = PP::OsdHide,
-                                        .showScaleFrom = 0.8,
-                                        .hideScaleTo = 0.9,
-                                        .showShaderEffectId = resolveShaderEffect(tree, PP::OsdShow),
-                                        .hideShaderEffectId = resolveShaderEffect(tree, PP::OsdHide),
-                                        .showShaderProfile = PP::OsdShow,
-                                        .hideShaderProfile = PP::OsdHide,
-                                        .showShaderParameters = resolveShaderParameters(tree, PP::OsdShow),
-                                        .hideShaderParameters = resolveShaderParameters(tree, PP::OsdHide)};
-}
-
-/// LayoutPicker: OSD-style fade-and-pop shape with a softer scale
-/// envelope (0.9→1 vs the OSD's 0.8→1) since the picker is a larger
-/// surface.
-///
-/// **Popup surface family — dedicated path partition.** Every leg
-/// resolves under `popup.layoutPicker.*`, NOT under `osd.*`. A
-/// Settings-UI edit at `popup.layoutPicker.hide` affects ONLY
-/// the layout picker; OSD timings stay independent. With no override
-/// set, `resolveWithInheritance` walks up to `popup` and finally
-/// to library defaults (150 ms OutCubic).
-///
-/// Shader legs key on the same `.show` / `.hide` leaves so a user can
-/// dissolve in and slide out (or any asymmetric pair). Both leaves walk
-/// up to `popup.layoutPicker` and on to `popup`, so a user
-/// who wants symmetric treatment overrides the surface path once and
-/// skips the leaves.
-PAL::SurfaceAnimator::Config buildLayoutPickerConfig(const PAS::ShaderProfileTree& tree)
-{
-    namespace PP = PhosphorAnimation::ProfilePaths;
-    return PAL::SurfaceAnimator::Config{
-        .showProfile = PP::PopupLayoutPickerShow,
-        .hideProfile = PP::PopupLayoutPickerHide,
-        .showScaleProfile = PP::PopupLayoutPickerShow,
-        .hideScaleProfile = PP::PopupLayoutPickerHide,
-        .showScaleFrom = 0.9,
-        .hideScaleTo = 0.95,
-        .showShaderEffectId = resolveShaderEffect(tree, PP::PopupLayoutPickerShow),
-        .hideShaderEffectId = resolveShaderEffect(tree, PP::PopupLayoutPickerHide),
-        .showShaderProfile = PP::PopupLayoutPickerShow,
-        .hideShaderProfile = PP::PopupLayoutPickerHide,
-        .showShaderParameters = resolveShaderParameters(tree, PP::PopupLayoutPickerShow),
-        .hideShaderParameters = resolveShaderParameters(tree, PP::PopupLayoutPickerHide)};
-}
-
-/// ZoneSelector: opacity-only show/hide. `keepMappedOnHide=true` so the
-/// hide animation actually paints.
-///
-/// **Popup surface family — dedicated path partition.** Every leg
-/// resolves under `popup.zoneSelector.*`, NOT under the shared
-/// `popup` baseline or the generic `widget.fadeOut` it previously
-/// borrowed. A Settings-UI edit at `popup.zoneSelector.hide`
-/// affects ONLY the zone selector. With no override set,
-/// `resolveWithInheritance` walks up to `popup` then library
-/// defaults.
-PAL::SurfaceAnimator::Config buildZoneSelectorConfig(const PAS::ShaderProfileTree& tree)
-{
-    namespace PP = PhosphorAnimation::ProfilePaths;
-    return PAL::SurfaceAnimator::Config{
-        .showProfile = PP::PopupZoneSelectorShow,
-        .hideProfile = PP::PopupZoneSelectorHide,
-        .showScaleProfile = {},
-        .hideScaleProfile = {},
-        .showShaderEffectId = resolveShaderEffect(tree, PP::PopupZoneSelectorShow),
-        .hideShaderEffectId = resolveShaderEffect(tree, PP::PopupZoneSelectorHide),
-        .showShaderProfile = PP::PopupZoneSelectorShow,
-        .hideShaderProfile = PP::PopupZoneSelectorHide,
-        .showShaderParameters = resolveShaderParameters(tree, PP::PopupZoneSelectorShow),
-        .hideShaderParameters = resolveShaderParameters(tree, PP::PopupZoneSelectorHide)};
-}
-
-/// SnapAssist: pop-in show only. The overlay uses destroy-on-hide
-/// (keepMappedOnHide=false), so ~Surface synchronously cancels any
-/// in-flight beginHide before the hide animation can paint a frame.
-/// Only `popup.snapAssist.show` is meaningful — `.hide` is
-/// intentionally absent from the taxonomy because no frame ever paints.
-///
-/// **Inheritance caveat (popup-family-only).** ShaderProfileTree::resolve
-/// walks parent paths, so within the popup family a user setting
-/// `popup` to dissolve cascades to LayoutPicker and ZoneSelector
-/// (both legs) and to SnapAssist's show leg via the chain
-/// `popup.<surface>.<leg>` → `popup.<surface>` →
-/// `popup` → `global`. SnapAssist's HIDE leg deliberately drops
-/// that walk-up — the surface destroys before paint, so the resolve
-/// would be contract-pure-and-runtime-no-op. The genuine OSD
-/// (`osd.show`/`osd.hide`) is in a separate subtree and is NOT touched
-/// by `popup` overrides regardless. If someone ever flips
-/// `keepMappedOnHide` to true on snap-assist, restore the
-/// `resolveShaderEffect(tree, PopupSnapAssistHide)` line below in
-/// lockstep with adding the new path constant.
-PAL::SurfaceAnimator::Config buildSnapAssistConfig(const PAS::ShaderProfileTree& tree)
-{
-    namespace PP = PhosphorAnimation::ProfilePaths;
-    return PAL::SurfaceAnimator::Config{// Popup surface family — dedicated path. A user editing
-                                        // `popup.snapAssist.show.json` affects ONLY the snap
-                                        // assist; siblings are unaffected. Built-in default mirrors the
-                                        // prior `popup` (150 ms widget-out) so behaviour is
-                                        // preserved.
-                                        .showProfile = PP::PopupSnapAssistShow,
-                                        .hideProfile = {},
-                                        .showScaleProfile = {},
-                                        .hideScaleProfile = {},
-                                        .showShaderEffectId = resolveShaderEffect(tree, PP::PopupSnapAssistShow),
-                                        .hideShaderEffectId = {},
-                                        .showShaderProfile = PP::PopupSnapAssistShow,
-                                        .hideShaderProfile = {},
-                                        .showShaderParameters = resolveShaderParameters(tree, PP::PopupSnapAssistShow),
-                                        .hideShaderParameters = {}};
-}
-
-} // namespace
-
-void OverlayService::setupSurfaceAnimator(PhosphorAnimation::PhosphorProfileRegistry& profileRegistry)
-{
-    namespace PAL = PhosphorAnimationLayer;
-
-    // Two existing surface types deliberately BYPASS this animator and
-    // therefore never read any config (default or per-role):
-    //   - Overlay (zone overlay rendering): hidden via direct
-    //     window->hide() in dismissOverlayWindow because the rapid
-    //     drag-mode path uses _idled property toggling rather than the
-    //     show/hide state machine.
-    //   - ShaderPreview (editor preview window): shown via direct
-    //     window->show() in showShaderPreview because the editor controls
-    //     visibility imperatively and re-creates on every open.
-    m_surfaceAnimator = std::make_unique<PAL::SurfaceAnimator>(profileRegistry, buildDefaultConfig());
-    if (m_animShaderRegistry) {
-        m_surfaceAnimator->setAnimationShaderRegistry(m_animShaderRegistry);
-    }
-
-    // Profile names are the same paths PhosphorMotionAnimation in QML
-    // uses today, so the live-reload path (drop a JSON, see it apply on
-    // next show) automatically applies to the C++ side too.
-    //
-    // Phase-2 surface unification: a single per-screen
-    // PzRoles::Notification surface (NotificationOverlay.qml) hosts both
-    // layout-OSD and navigation-OSD content via a Loader. Production
-    // surfaces are scoped `plasmazones-notification-{screenId}-{gen}`
-    // and resolve to this config via the animator's longest-prefix
-    // lookup.
-    //
-    // Initial registration runs with an empty tree — m_settings is wired
-    // later via setSettings(). A default-constructed tree resolves every
-    // path to an empty effect id, so this pass installs motion-only
-    // configs (identical to the pre-shader-wireup behaviour). Once
-    // settings exist, setSettings calls applyShaderProfilesToAnimator
-    // again with the live tree, and connects shaderProfileTreeChanged
-    // for live-reload. This keeps the constructor's invariant ("animator
-    // is ready before any Surface is created") while deferring the
-    // shader wiring to the moment settings are available.
-    applyShaderProfilesToAnimator(PAS::ShaderProfileTree{});
-}
-
-void OverlayService::applyShaderProfilesToAnimator(const PAS::ShaderProfileTree& tree)
-{
-    if (!m_surfaceAnimator) {
-        return;
-    }
-    // Diagnostic log gated on lcOverlay().isDebugEnabled() — qCDebug
-    // gates the OUTPUT but Qt evaluates argument expressions
-    // unconditionally, so the seven extra resolveShaderEffect calls
-    // would run even when debug logging is disabled. Each
-    // ShaderProfileTree::resolve walks the parent chain and overlay-
-    // merges every override; on a typical settings-edit signal storm
-    // this is wasted work the explicit gate eliminates.
-    if (lcOverlay().isDebugEnabled()) {
-        namespace PP = PhosphorAnimation::ProfilePaths;
-        qCDebug(lcOverlay).nospace() << "applyShaderProfilesToAnimator — overrides=" << tree.overriddenPaths().size()
-                                     << " resolved: osd.show=" << resolveShaderEffect(tree, PP::OsdShow)
-                                     << " osd.hide=" << resolveShaderEffect(tree, PP::OsdHide)
-                                     << " zoneSelector.show=" << resolveShaderEffect(tree, PP::PopupZoneSelectorShow)
-                                     << " zoneSelector.hide=" << resolveShaderEffect(tree, PP::PopupZoneSelectorHide)
-                                     << " layoutPicker.show=" << resolveShaderEffect(tree, PP::PopupLayoutPickerShow)
-                                     << " layoutPicker.hide=" << resolveShaderEffect(tree, PP::PopupLayoutPickerHide)
-                                     << " snapAssist.show=" << resolveShaderEffect(tree, PP::PopupSnapAssistShow);
-    }
-    m_surfaceAnimator->registerConfigForRole(PzRoles::Notification, buildOsdConfig(tree));
-    m_surfaceAnimator->registerConfigForRole(PzRoles::LayoutPicker, buildLayoutPickerConfig(tree));
-    m_surfaceAnimator->registerConfigForRole(PzRoles::ZoneSelector, buildZoneSelectorConfig(tree));
-    m_surfaceAnimator->registerConfigForRole(PzRoles::SnapAssist, buildSnapAssistConfig(tree));
-}
+// Per-role SurfaceAnimator config builders + setupSurfaceAnimator +
+// applyShaderProfilesToAnimator are extracted to
+// overlayservice/animation_config.cpp to keep this translation unit
+// under the project's <800-line guideline.
 
 void OverlayService::primeSurfaceRenderPipeline(PhosphorLayer::Surface* surface)
 {
@@ -915,251 +639,9 @@ PhosphorLayer::Surface* OverlayService::createWarmedOsdSurface(const PhosphorLay
     return surface;
 }
 
-void OverlayService::show()
-{
-    if (m_visible) {
-        return;
-    }
-
-    // Check if we should show on all monitors or just the cursor's screen
-    bool showOnAllMonitors = !m_settings || m_settings->showZonesOnAllMonitors();
-
-    QScreen* cursorScreen = nullptr;
-    if (!showOnAllMonitors) {
-        // Find the screen containing the cursor
-        cursorScreen = QGuiApplication::screenAt(QCursor::pos());
-        if (!cursorScreen) {
-            // Fallback to primary screen if cursor position detection fails
-            cursorScreen = Utils::primaryScreen();
-        }
-        // If the cursor's screen has PlasmaZones disabled, don't show overlay at all
-        // Check both physical and effective (virtual) screen IDs
-        if (cursorScreen && m_settings) {
-            QString effectiveId = Utils::effectiveScreenIdAt(m_screenManager, QCursor::pos(), cursorScreen);
-            if (isContextDisabled(m_settings, PhosphorZones::AssignmentEntry::Snapping, effectiveId,
-                                  m_currentVirtualDesktop, m_currentActivity)) {
-                return;
-            }
-        }
-    }
-
-    initializeOverlay(cursorScreen);
-}
-
-void OverlayService::showAtPosition(int cursorX, int cursorY)
-{
-    // Check if we should show on all monitors or just the cursor's screen
-    bool showOnAllMonitors = !m_settings || m_settings->showZonesOnAllMonitors();
-
-    QScreen* cursorScreen = nullptr;
-    if (!showOnAllMonitors) {
-        // Find the screen containing the cursor using provided coordinates
-        // This works on Wayland where QCursor::pos() doesn't work
-        cursorScreen = Utils::findScreenAtPosition(cursorX, cursorY);
-        if (!cursorScreen) {
-            // Fallback to primary screen if no screen contains the cursor position
-            cursorScreen = Utils::primaryScreen();
-        }
-
-        // If the cursor's screen has PlasmaZones disabled, don't show overlay at all
-        // Check both physical and effective (virtual) screen IDs
-        if (cursorScreen && m_settings) {
-            QString effectiveId = Utils::effectiveScreenIdAt(m_screenManager, QPoint(cursorX, cursorY), cursorScreen);
-            if (isContextDisabled(m_settings, PhosphorZones::AssignmentEntry::Snapping, effectiveId,
-                                  m_currentVirtualDesktop, m_currentActivity)) {
-                return;
-            }
-        }
-    }
-
-    const QPoint cursorPos(cursorX, cursorY);
-
-    if (m_visible) {
-        // One-overlay-per-VS architecture: every VS already has a live
-        // overlay window from initializeOverlay. Cross-VS switching is
-        // just a matter of flipping per-window _idled state — no
-        // re-init, no rekey, no layer-shell re-anchor. This sidesteps
-        // the earlier "wrong spot" bug where rekey moved the map entry
-        // but left the layer surface anchored to the previous VS's
-        // bounds, and the full NVIDIA vkDestroyDevice deadlock on any
-        // destroy path.
-        if (!cursorScreen) {
-            cursorScreen = Utils::findScreenAtPosition(cursorPos);
-        }
-        if (!cursorScreen) {
-            return;
-        }
-        const QString cursorEffectiveId =
-            Utils::effectiveScreenIdAt(m_screenManager, QPoint(cursorX, cursorY), cursorScreen);
-        if (cursorEffectiveId.isEmpty()) {
-            return;
-        }
-        // Per-VS toggle (autotile→snap mid-session) leaves us with a VS that
-        // is no longer excluded but never had its overlay window created at
-        // first-show time. Detect that here — the cursor is now on a non-
-        // excluded VS with no live window — and fall through to full init
-        // so Phase 3 creates the window. Without this, applyIdleStateForCursor
-        // finds nothing to flip and the overlay never becomes visible until
-        // the next hide()/show() cycle.
-        const bool cursorVsHasWindow = m_screenStates.contains(cursorEffectiveId)
-            && m_screenStates.value(cursorEffectiveId).overlayWindow != nullptr;
-        if (cursorVsHasWindow || m_excludedScreens.contains(cursorEffectiveId)) {
-            m_currentOverlayScreenId = showOnAllMonitors ? QString() : cursorEffectiveId;
-            applyIdleStateForCursor(cursorEffectiveId, showOnAllMonitors);
-            return;
-        }
-        // Fall through — initializeOverlay will (re)build the per-VS window
-        // set against the current excluded set and resume normal operation.
-    }
-
-    initializeOverlay(cursorScreen, cursorPos);
-}
-
-void OverlayService::hide()
-{
-    if (!m_visible) {
-        return;
-    }
-
-    m_visible = false;
-    m_currentOverlayScreenId.clear();
-
-    // Stop shader animation
-    stopShaderAnimation();
-
-    // Do NOT invalidate m_shaderTimer - keeps iTime continuous across show/hide
-    // so animations feel less predictable and don't restart
-
-    // Destroy overlay windows instead of hiding them. On Vulkan with Wayland
-    // layer-shell, window->hide() destroys the wl_surface but the Qt Vulkan
-    // backend doesn't properly reinitialize the VkSwapchainKHR when the window
-    // is re-shown, causing the scene graph render loop to stall. Destroying the
-    // window entirely and creating a fresh one on the next show() avoids this.
-    // initializeOverlay() will call createOverlayWindow() since overlay windows
-    // are now destroyed.
-    const QStringList screenIds = m_screenStates.keys();
-    for (const QString& screenId : screenIds) {
-        destroyOverlayWindow(screenId);
-    }
-
-    m_pendingShaderError.clear();
-
-    Q_EMIT visibilityChanged(false);
-}
-
-void OverlayService::toggle()
-{
-    if (m_visible) {
-        hide();
-    } else {
-        show();
-    }
-}
-
-void OverlayService::setIdleForDragPause()
-{
-    // Blank the overlay's shader output without destroying QQuickWindows.
-    // The heavy hide() path pays a ~QQuickWindow Vulkan teardown per screen
-    // which blocks the main thread on the scene graph render thread — and
-    // with modifier-key thrashing during a drag we ended up paying that cost
-    // many times per second, stalling D-Bus dispatch long enough for
-    // kwin-effect's endDrag to time out and the user to see multi-second lag.
-    //
-    // Here we only clear the per-window QML properties that drive the shader
-    // (zones, zoneCount, highlights). Windows, Vulkan swap chains, and layer
-    // surfaces stay alive. On the next activation tick, refreshFromIdle()
-    // re-pushes the current zone data — cheap because the labels-texture
-    // build is hash-cached on unchanged inputs.
-    if (!m_visible) {
-        return;
-    }
-    for (auto it = m_screenStates.begin(); it != m_screenStates.end(); ++it) {
-        QQuickWindow* window = it.value().overlayWindow;
-        if (!window) {
-            continue;
-        }
-        // _idled gates content.visible and toggles Qt.WindowTransparentForInput
-        // in the overlay QML (RenderNodeOverlay.qml / ZoneOverlay.qml). That
-        // makes the wl_surface effectively invisible and non-input-absorbing
-        // in place, without destroying the QQuickWindow. Blanking only the
-        // zones properties below is not sufficient — on some shaders the
-        // base pass still renders visible output when zoneCount==0, and the
-        // input region stays active until the flag change lands.
-        writeQmlProperty(window, QStringLiteral("_idled"), true);
-        writeQmlProperty(window, QStringLiteral("zones"), QVariantList());
-        writeQmlProperty(window, QStringLiteral("zoneCount"), 0);
-        writeQmlProperty(window, QStringLiteral("highlightedCount"), 0);
-        writeQmlProperty(window, QStringLiteral("highlightedZoneId"), QString());
-        writeQmlProperty(window, QStringLiteral("highlightedZoneIds"), QVariantList());
-        // NOTE: labelsTextureHash is intentionally NOT cleared here. The QML
-        // side's labelsTexture property still holds the previously-built image
-        // (setProperty was never called with a new one); it just isn't sampled
-        // while zoneCount is 0. Keeping the hash means refreshFromIdle() with
-        // unchanged zones hits the cache and costs one hash compute instead
-        // of rebuilding 23 MB of pixels.
-    }
-    // CRITICAL: mark zone data CLEAN, not dirty. The shader animation tick
-    // (shader.cpp:245) re-runs updateZonesForAllWindows() whenever dirty is
-    // set, which would rebuild the real zones and undo the blank. The idle
-    // state is "what we just wrote, do not re-derive from layout data until
-    // refreshFromIdle() is called."
-    m_zoneDataDirty = false;
-    // NOTE: we deliberately do NOT call stopShaderAnimation() here. The
-    // shader timer keeps ticking at ~60 Hz while idled, but with zoneCount
-    // set to 0 the per-frame work collapses to a handful of uniform uploads
-    // to a surface that's rendering no visible geometry — bounded cost, O(1)
-    // per screen. Pausing and restarting the timer across the idle cycle
-    // would require additional state tracking in refreshFromIdle() and add
-    // a startup transient on every modifier re-press. Left unchanged for
-    // simplicity; revisit if profiling ever shows it as a hot spot.
-}
-
-void OverlayService::refreshFromIdle()
-{
-    // Restore zone data after a setIdleForDragPause() blank and flip
-    // the active VS's overlay back to visible.
-    //
-    // setIdleForDragPause() unconditionally idles every overlay (zones
-    // blanked + _idled=true), so refreshFromIdle() re-pushes zone data
-    // to all of them and then applies the cursor-based idle state to
-    // un-idle the one the cursor is currently on. The L2 labels-texture
-    // hash cache keeps the shader-path re-push cheap on unchanged inputs.
-    if (!m_visible) {
-        return;
-    }
-    updateZonesForAllWindows();
-    // Resolve the cursor's current VS — the drag adaptor keeps
-    // m_currentOverlayScreenId updated via showAtPosition, so this
-    // reflects the last VS the cursor was observed on.
-    const bool showOnAllMonitors = !m_settings || m_settings->showZonesOnAllMonitors();
-    applyIdleStateForCursor(m_currentOverlayScreenId, showOnAllMonitors);
-}
-
-void OverlayService::applyIdleStateForCursor(const QString& activeEffectiveId, bool showOnAllMonitors)
-{
-    // One-overlay-per-VS idle state: iterate every live overlay window
-    // and flip its _idled QML property based on whether its VS should
-    // currently be accepting input / rendering content.
-    //
-    // - showOnAllMonitors=true  → all overlays un-idled (all VSes active)
-    // - showOnAllMonitors=false → only activeEffectiveId un-idled
-    // - activeEffectiveId empty → all overlays idled (no active VS —
-    //   used by setIdleForDragPause when drag-end hasn't chosen a next
-    //   cursor position yet, or when the cursor sits on a disabled VS)
-    //
-    // The write is idempotent: QML property binding only re-evaluates
-    // when the value actually changes, so flipping _idled on a window
-    // that's already in the target state is free.
-    for (auto it = m_screenStates.begin(); it != m_screenStates.end(); ++it) {
-        QQuickWindow* window = it.value().overlayWindow;
-        if (!window) {
-            continue;
-        }
-        const bool shouldBeActive =
-            showOnAllMonitors || (it.key() == activeEffectiveId && !activeEffectiveId.isEmpty());
-        writeQmlProperty(window, QStringLiteral("_idled"), !shouldBeActive);
-    }
-}
+// Overlay show/hide/toggle + setIdleForDragPause/refreshFromIdle/
+// applyIdleStateForCursor are extracted to overlayservice/lifecycle.cpp
+// alongside the existing selector/snapassist/osd splits.
 
 void OverlayService::setAnimationShaderRegistry(PhosphorAnimationShaders::AnimationShaderRegistry* registry)
 {
