+++ src/editor/qml/EditorWindow.qml
@@ -349,7 +349,11 @@ Window {
 
         Behavior on opacity {
             PhosphorMotionAnimation {
-                profile: opacity > 0.5 ? "widget.fadeIn" : "widget.fadeOut"
+                // Direction is taken from the visibility condition so the leg
+                // is decided synchronously when fullscreen toggles. Reading
+                // the animated `opacity` would re-evaluate during the
+                // Behavior and flip the leg mid-animation.
+                profile: !editorWindow.fullscreenMode ? "widget.fadeIn" : "widget.fadeOut"
                 durationOverride: 200
             }
 
@@ -735,7 +739,10 @@ Window {
 
         Behavior on opacity {
             PhosphorMotionAnimation {
-                profile: opacity > 0.5 ? "widget.fadeIn" : "widget.fadeOut"
+                // Direction is taken from the same condition driving the
+                // `opacity` binding above (`fullscreenMode ? 1 : 0`) so the
+                // leg is decided synchronously when fullscreen toggles.
+                profile: editorWindow.fullscreenMode ? "widget.fadeIn" : "widget.fadeOut"
                 durationOverride: 200
             }
