+++ src/settings/qml/AnimationsPopupsPage.qml
@@ -8,8 +8,12 @@ import org.kde.kirigami as Kirigami
 // Popups animation page — the `popup.*` resolver subtree. The "All
 // Popups" parent is a parent-node card whose override cascades to
 // `popup.zoneSelector.*`, `popup.layoutPicker.*` and
-// `popup.snapAssist.*` via ShaderProfileTree::resolve's walk-up
+// `popup.snapAssist.show` via ShaderProfileTree::resolve's walk-up
 // (`popup` is the closest common ancestor of all three popup surfaces).
+// SnapAssist's hide leg is intentionally omitted — the surface
+// destroys on hide before any frame paints, so a profile assignment
+// there would be runtime-dead (see `animationshadersupportedpaths.h`
+// and `OverlayService::buildSnapAssistConfig`).
 // In-app side panels (settings nav rail, editor property panel) live
 // under `panel.*` and have their own dedicated Panels page.
 Flickable {
@@ -58,12 +62,6 @@ Flickable {
             Layout.fillWidth: true
             eventPath: "popup.snapAssist.show"
             eventLabel: i18n("Snap Assist — Show")
-        }
-
-        AnimationEventCard {
-            Layout.fillWidth: true
-            eventPath: "popup.snapAssist.hide"
-            eventLabel: i18n("Snap Assist — Hide")
         }
 
     }
