+++ tests/unit/settings/test_animations_page_controller.cpp
@@ -3,7 +3,7 @@
 
 /**
  * @file test_animations_page_controller.cpp
- * @brief Phase 0 round-trip + path-discovery tests for AnimationsPageController.
+ * @brief Path-discovery + override-CRUD tests for AnimationsPageController.
  *
  * Pins the file-per-path persistence model: setOverride writes one JSON
  * file under `<userProfilesDir>/<path>.json`, clearOverride deletes it,
@@ -11,6 +11,10 @@
  *
  * Uses `setUserProfilesDirOverride()` to redirect file I/O into a
  * tmpdir so the test never touches the real user XDG dir.
+ *
+ * Companion test files (split for the <800-line guideline):
+ *   - test_animations_motion_sets.cpp      — preset / motion-set / pending
+ *   - test_animations_shader_overrides.cpp — shader-effect overrides
  */
 
 #include <QSignalSpy>
@@ -20,21 +24,15 @@
 #include <QDir>
 #include <QFile>
 #include <QFileInfo>
-#include <QJsonArray>
 #include <QJsonDocument>
 #include <QJsonObject>
-#include <QRegularExpression>
 
 #include <PhosphorAnimation/Profile.h>
 #include <PhosphorAnimation/ProfilePaths.h>
 
-#include "config/settings.h"
 #include "settings/animationspagecontroller.h"
-#include "../helpers/IsolatedConfigGuard.h"
-#include <PhosphorAnimation/AnimationShaderRegistry.h>
 
 using namespace PlasmaZones;
-using PlasmaZones::TestHelpers::IsolatedConfigGuard;
 
 namespace {
 
@@ -316,343 +314,6 @@ private Q_SLOTS:
         QCOMPARE(resolved.value(QStringLiteral("duration")).toInt(), 333);
     }
 
-    // ─── User preset library ──────────────────────────────────────────────
-
-    void addUserPreset_writesFileWithSlugFilename()
-    {
-        QTemporaryDir tmp;
-        QVERIFY(tmp.isValid());
-        AnimationsPageController c;
-        c.setUserProfilesDirOverride(tmp.path());
-
-        QSignalSpy spy(&c, &AnimationsPageController::userPresetsChanged);
-        QVERIFY(c.addUserPreset(
-            QStringLiteral("My Snappy Spring!"),
-            {{QStringLiteral("curve"), QStringLiteral("spring:14.0,0.7")}, {QStringLiteral("duration"), 200}}));
-        QCOMPARE(spy.count(), 1);
-
-        // Slugified filename: lowercase, non-alnum collapsed to '-',
-        // trailing '-' trimmed.
-        QVERIFY(QFileInfo::exists(tmp.path() + QStringLiteral("/my-snappy-spring.json")));
-    }
-
-    void userPresets_excludesPathNamedFiles()
-    {
-        QTemporaryDir tmp;
-        QVERIFY(tmp.isValid());
-        AnimationsPageController c;
-        c.setUserProfilesDirOverride(tmp.path());
-
-        // Preset file
-        QVERIFY(
-            c.addUserPreset(QStringLiteral("My Curve"), {{QStringLiteral("curve"), QStringLiteral("0.5,0,0.5,1")}}));
-        // Override file at a known path
-        QVERIFY(c.setOverride(QStringLiteral("zone.snapIn"), {{QStringLiteral("duration"), 250}}));
-
-        // userPresets sees ONLY the preset, not the override
-        const QVariantList presets = c.userPresets();
-        QCOMPARE(presets.size(), 1);
-        QCOMPARE(presets.first().toMap().value(QStringLiteral("name")).toString(), QStringLiteral("My Curve"));
-    }
-
-    void addUserPreset_rejectsKnownPathNames()
-    {
-        QTemporaryDir tmp;
-        QVERIFY(tmp.isValid());
-        AnimationsPageController c;
-        c.setUserProfilesDirOverride(tmp.path());
-
-        QSignalSpy spy(&c, &AnimationsPageController::userPresetsChanged);
-        // "zone.snapIn" is a known event path — would shadow the
-        // override slot.
-        QVERIFY(!c.addUserPreset(QStringLiteral("zone.snapIn"), {{QStringLiteral("duration"), 100}}));
-        QCOMPARE(spy.count(), 0);
-        QVERIFY(c.userPresets().isEmpty());
-    }
-
-    void addUserPreset_rejectsEmptyAndAllSymbol()
-    {
-        QTemporaryDir tmp;
-        QVERIFY(tmp.isValid());
-        AnimationsPageController c;
-        c.setUserProfilesDirOverride(tmp.path());
-
-        QVERIFY(!c.addUserPreset(QString(), {{QStringLiteral("duration"), 100}}));
-        // All-symbol slugifies to empty → reject (would write to ".json")
-        QVERIFY(!c.addUserPreset(QStringLiteral("@@@"), {{QStringLiteral("duration"), 100}}));
-    }
-
-    void removeUserPreset_emitsAndDeletes()
-    {
-        QTemporaryDir tmp;
-        QVERIFY(tmp.isValid());
-        AnimationsPageController c;
-        c.setUserProfilesDirOverride(tmp.path());
-
-        QVERIFY(
-            c.addUserPreset(QStringLiteral("To Delete"), {{QStringLiteral("curve"), QStringLiteral("spring:10,0.5")}}));
-        QCOMPARE(c.userPresets().size(), 1);
-
-        QSignalSpy spy(&c, &AnimationsPageController::userPresetsChanged);
-        QVERIFY(c.removeUserPreset(QStringLiteral("To Delete")));
-        QCOMPARE(spy.count(), 1);
-        QVERIFY(c.userPresets().isEmpty());
-    }
-
-    void removeUserPreset_unknownReturnsFalse()
-    {
-        QTemporaryDir tmp;
-        QVERIFY(tmp.isValid());
-        AnimationsPageController c;
-        c.setUserProfilesDirOverride(tmp.path());
-
-        QSignalSpy spy(&c, &AnimationsPageController::userPresetsChanged);
-        QVERIFY(!c.removeUserPreset(QStringLiteral("nonexistent")));
-        QCOMPARE(spy.count(), 0);
-    }
-
-    // ─── Motion sets ──────────────────────────────────────────────────────
-
-    void saveCurrentAsMotionSet_capturesPathOverridesOnly()
-    {
-        QTemporaryDir tmp;
-        QVERIFY(tmp.isValid());
-        AnimationsPageController c;
-        c.setUserProfilesDirOverride(tmp.path());
-
-        // Mix of path overrides and a user preset
-        QVERIFY(c.setOverride(QStringLiteral("zone.snapIn"), {{QStringLiteral("duration"), 222}}));
-        QVERIFY(
-            c.setOverride(QStringLiteral("osd.show"), {{QStringLiteral("curve"), QStringLiteral("spring:10,0.7")}}));
-        QVERIFY(
-            c.addUserPreset(QStringLiteral("My Preset"), {{QStringLiteral("curve"), QStringLiteral("0.5,0,0.5,1")}}));
-
-        QSignalSpy spy(&c, &AnimationsPageController::motionSetsChanged);
-        QVERIFY(c.saveCurrentAsMotionSet(QStringLiteral("My Set"), QStringLiteral("test set")));
-        QCOMPARE(spy.count(), 1);
-
-        const QVariantList sets = c.availableMotionSets();
-        QCOMPARE(sets.size(), 1);
-        const QVariantMap set = sets.first().toMap();
-        QCOMPARE(set.value(QStringLiteral("name")).toString(), QStringLiteral("My Set"));
-        // Should capture the 2 path overrides, NOT the preset
-        QCOMPARE(set.value(QStringLiteral("overrideCount")).toInt(), 2);
-    }
-
-    void applyMotionSet_writesPerPathFiles()
-    {
-        QTemporaryDir tmp;
-        QVERIFY(tmp.isValid());
-        AnimationsPageController c;
-        c.setUserProfilesDirOverride(tmp.path());
-
-        // Build set, then clear overrides, then apply
-        QVERIFY(c.setOverride(QStringLiteral("zone.snapIn"), {{QStringLiteral("duration"), 333}}));
-        QVERIFY(c.saveCurrentAsMotionSet(QStringLiteral("snappy-set"), QString()));
-        QVERIFY(c.clearOverride(QStringLiteral("zone.snapIn")));
-        QVERIFY(!c.hasOverride(QStringLiteral("zone.snapIn")));
-
-        QVERIFY(c.applyMotionSet(QStringLiteral("snappy-set")));
-        QVERIFY(c.hasOverride(QStringLiteral("zone.snapIn")));
-        QCOMPARE(c.rawProfile(QStringLiteral("zone.snapIn")).value(QStringLiteral("duration")).toInt(), 333);
-    }
-
-    void applyMotionSet_mergesPreservesOtherPaths()
-    {
-        QTemporaryDir tmp;
-        QVERIFY(tmp.isValid());
-        AnimationsPageController c;
-        c.setUserProfilesDirOverride(tmp.path());
-
-        // Save a set with one path
-        QVERIFY(c.setOverride(QStringLiteral("zone.snapIn"), {{QStringLiteral("duration"), 222}}));
-        QVERIFY(c.saveCurrentAsMotionSet(QStringLiteral("set-a"), QString()));
-        QVERIFY(c.clearOverride(QStringLiteral("zone.snapIn")));
-
-        // Set an UNRELATED override
-        QVERIFY(c.setOverride(QStringLiteral("osd.show"), {{QStringLiteral("duration"), 555}}));
-
-        // Apply set-a; osd.show should still be 555 (merge, not replace)
-        QVERIFY(c.applyMotionSet(QStringLiteral("set-a")));
-        QCOMPARE(c.rawProfile(QStringLiteral("osd.show")).value(QStringLiteral("duration")).toInt(), 555);
-        QVERIFY(c.hasOverride(QStringLiteral("zone.snapIn")));
-    }
-
-    void removeMotionSet_emitsAndDeletes()
-    {
-        QTemporaryDir tmp;
-        QVERIFY(tmp.isValid());
-        AnimationsPageController c;
-        c.setUserProfilesDirOverride(tmp.path());
-
-        QVERIFY(c.setOverride(QStringLiteral("zone.snapIn"), {{QStringLiteral("duration"), 222}}));
-        QVERIFY(c.saveCurrentAsMotionSet(QStringLiteral("To Remove"), QString()));
-        QCOMPARE(c.availableMotionSets().size(), 1);
-
-        QSignalSpy spy(&c, &AnimationsPageController::motionSetsChanged);
-        QVERIFY(c.removeMotionSet(QStringLiteral("To Remove")));
-        QCOMPARE(spy.count(), 1);
-        QVERIFY(c.availableMotionSets().isEmpty());
-    }
-
-    // ─── Pending changes / commit / revert ────────────────────────────────
-
-    void hasPendingChanges_falseInitially()
-    {
-        QTemporaryDir tmp;
-        QVERIFY(tmp.isValid());
-        AnimationsPageController c;
-        c.setUserProfilesDirOverride(tmp.path());
-        QVERIFY(!c.hasPendingChanges());
-    }
-
-    void setOverride_emitsPendingChangesChanged()
-    {
-        QTemporaryDir tmp;
-        QVERIFY(tmp.isValid());
-        AnimationsPageController c;
-        c.setUserProfilesDirOverride(tmp.path());
-
-        QSignalSpy spy(&c, &AnimationsPageController::pendingChangesChanged);
-        QVERIFY(c.setOverride(QStringLiteral("zone.snapIn"), {{QStringLiteral("duration"), 250}}));
-        QCOMPARE(spy.count(), 1);
-        QVERIFY(c.hasPendingChanges());
-    }
-
-    void revertPending_restoresPreEditFile()
-    {
-        QTemporaryDir tmp;
-        QVERIFY(tmp.isValid());
-        AnimationsPageController c;
-        c.setUserProfilesDirOverride(tmp.path());
-
-        // Establish a pre-edit baseline: write 100 ms.
-        QVERIFY(c.setOverride(QStringLiteral("zone.snapIn"), {{QStringLiteral("duration"), 100}}));
-        c.commitPending();
-        QVERIFY(!c.hasPendingChanges());
-
-        // Edit to 250 ms → should snapshot the 100ms file.
-        QVERIFY(c.setOverride(QStringLiteral("zone.snapIn"), {{QStringLiteral("duration"), 250}}));
-        QCOMPARE(c.rawProfile(QStringLiteral("zone.snapIn")).value(QStringLiteral("duration")).toInt(), 250);
-
-        // Revert → file content restored to 100 ms.
-        c.revertPending();
-        QVERIFY(!c.hasPendingChanges());
-        QCOMPARE(c.rawProfile(QStringLiteral("zone.snapIn")).value(QStringLiteral("duration")).toInt(), 100);
-    }
-
-    void revertPending_deletesFilesThatDidntExistBefore()
-    {
-        QTemporaryDir tmp;
-        QVERIFY(tmp.isValid());
-        AnimationsPageController c;
-        c.setUserProfilesDirOverride(tmp.path());
-
-        // Fresh override on a path with no prior file.
-        QVERIFY(c.setOverride(QStringLiteral("osd.show"), {{QStringLiteral("duration"), 200}}));
-        QVERIFY(c.hasOverride(QStringLiteral("osd.show")));
-
-        c.revertPending();
-        QVERIFY(!c.hasOverride(QStringLiteral("osd.show")));
-    }
-
-    void revertPending_emitsOverrideChangedPerPath()
-    {
-        QTemporaryDir tmp;
-        QVERIFY(tmp.isValid());
-        AnimationsPageController c;
-        c.setUserProfilesDirOverride(tmp.path());
-
-        QVERIFY(c.setOverride(QStringLiteral("zone.snapIn"), {{QStringLiteral("duration"), 200}}));
-        QVERIFY(c.setOverride(QStringLiteral("osd.show"), {{QStringLiteral("duration"), 300}}));
-
-        QSignalSpy spy(&c, &AnimationsPageController::overrideChanged);
-        c.revertPending();
-
-        // Two paths reverted → two emissions (one per file). Order is
-        // hash-iteration so we check membership rather than position.
-        QCOMPARE(spy.count(), 2);
-        QStringList emitted;
-        for (const auto& args : spy)
-            emitted << args.at(0).toString();
-        QVERIFY(emitted.contains(QStringLiteral("zone.snapIn")));
-        QVERIFY(emitted.contains(QStringLiteral("osd.show")));
-    }
-
-    void commitPending_clearsSnapshotWithoutEmittingDataChanged()
-    {
-        QTemporaryDir tmp;
-        QVERIFY(tmp.isValid());
-        AnimationsPageController c;
-        c.setUserProfilesDirOverride(tmp.path());
-
-        QVERIFY(c.setOverride(QStringLiteral("zone.snapIn"), {{QStringLiteral("duration"), 250}}));
-        QVERIFY(c.hasPendingChanges());
-
-        // commitPending clears the snapshot and only emits
-        // pendingChangesChanged (once). It must NOT re-fire the
-        // per-path overrideChanged or the data signals — the user just
-        // saved, no rows visually moved.
-        QSignalSpy pendingSpy(&c, &AnimationsPageController::pendingChangesChanged);
-        QSignalSpy overrideSpy(&c, &AnimationsPageController::overrideChanged);
-        c.commitPending();
-        QVERIFY(!c.hasPendingChanges());
-        QCOMPARE(pendingSpy.count(), 1);
-        QCOMPARE(overrideSpy.count(), 0);
-    }
-
-    // ─── Shader override end-to-end (the actual user flow) ──────────────
-
-    /// Simulates the full AnimationEventCard combo flow:
-    ///   user clicks pixelate
-    ///     → setShaderOverride(path, "pixelate", {})
-    ///   refresh fires
-    ///     → resolvedShaderProfile(path) → expect effectId="pixelate"
-    /// If resolvedShaderProfile returns empty effectId here, the QML
-    /// combo would snap back to "None" — mirrors the user's symptom.
-    void setShaderOverride_resolveReturnsTheJustWrittenEffect()
-    {
-        IsolatedConfigGuard guard;
-        Settings settings;
-        PhosphorAnimationShaders::AnimationShaderRegistry registry;
-        AnimationsPageController c(&registry, &settings);
-
-        const QString path = QStringLiteral("osd.show");
-
-        QVERIFY(c.setShaderOverride(path, QStringLiteral("pixelate"), {}));
-
-        const QVariantMap raw = c.rawShaderProfile(path);
-        const QVariantMap resolved = c.resolvedShaderProfile(path);
-
-        // Raw read MUST contain effectId=pixelate.
-        QCOMPARE(raw.value(QStringLiteral("effectId")).toString(), QStringLiteral("pixelate"));
-        // Resolved read (what the QML refresh path uses) MUST too.
-        QCOMPARE(resolved.value(QStringLiteral("effectId")).toString(), QStringLiteral("pixelate"));
-    }
-
-    /// Symptom-mirroring test: pick A, then immediately pick B.
-    /// Both reads must return the corresponding effect.
-    void setShaderOverride_repeatedPicksPreserveLatest()
-    {
-        IsolatedConfigGuard guard;
-        Settings settings;
-        PhosphorAnimationShaders::AnimationShaderRegistry registry;
-        AnimationsPageController c(&registry, &settings);
-
-        const QString path = QStringLiteral("osd.show");
-
-        QVERIFY(c.setShaderOverride(path, QStringLiteral("pixelate"), {}));
-        QCOMPARE(c.resolvedShaderProfile(path).value(QStringLiteral("effectId")).toString(),
-                 QStringLiteral("pixelate"));
-
-        QVERIFY(c.setShaderOverride(path, QStringLiteral("dissolve"), {}));
-        QCOMPARE(c.resolvedShaderProfile(path).value(QStringLiteral("effectId")).toString(),
-                 QStringLiteral("dissolve"));
-
-        QVERIFY(c.setShaderOverride(path, QStringLiteral("glitch"), {}));
-        QCOMPARE(c.resolvedShaderProfile(path).value(QStringLiteral("effectId")).toString(), QStringLiteral("glitch"));
-    }
-
     void resolvedProfile_partialLeafFillsFromParentAndDefaults()
     {
         QTemporaryDir tmp;
@@ -720,145 +381,32 @@ private Q_SLOTS:
         QVERIFY(!c.clearOverride(QStringLiteral("not.a.real.path")));
     }
 
-    /// `removeUserPreset` MUST not delete an override file even when its
-    /// embedded `name` field happens to match the supplied preset name.
-    /// Pre-fix, the directory-scan fallback walked every JSON file in
-    /// the profiles dir and matched by `name`; an override file whose
-    /// `name` matched the searched preset would be deleted.
-    void removeUserPreset_doesNotTouchOverrideFiles()
+    // ─── Pass-1 behaviour pins ────────────────────────────────────────────
+
+    /// `setOverride` compare-and-skip: writing the same Profile twice
+    /// must not fire a second `overrideChanged` / `pendingChangesChanged`.
+    /// Pre-fix the second write rewrote an identical file and lit Save.
+    void setOverride_repeatedWriteIsNoOp()
     {
         QTemporaryDir tmp;
         QVERIFY(tmp.isValid());
         AnimationsPageController c;
         c.setUserProfilesDirOverride(tmp.path());
 
-        // Write an override file. Its on-disk `name` field is "zone.snapIn"
-        // (per setOverride's stamping rule).
-        QVERIFY(c.setOverride(QStringLiteral("zone.snapIn"), {{QStringLiteral("duration"), 250}}));
-        const QString overrideFilePath = tmp.path() + QStringLiteral("/zone.snapIn.json");
-        QVERIFY(QFileInfo::exists(overrideFilePath));
+        const auto profile = QVariantMap{{QStringLiteral("duration"), 250},
+                                         {QStringLiteral("curve"), QStringLiteral("0.33,1.00,0.68,1.00")}};
 
-        // Hand-craft a hostile preset on disk: `name` field = "zone.snapIn"
-        // but the FILE is not at the canonical override slot. The library
-        // skips it on userPresets() (collision filter) but a naive
-        // remove-by-name walk would still delete it. We're asserting the
-        // override file in the canonical slot remains intact regardless.
-        QSignalSpy spy(&c, &AnimationsPageController::userPresetsChanged);
-        QVERIFY(!c.removeUserPreset(QStringLiteral("zone.snapIn")));
-        QCOMPARE(spy.count(), 0);
+        QSignalSpy overrideSpy(&c, &AnimationsPageController::overrideChanged);
+        QSignalSpy pendingSpy(&c, &AnimationsPageController::pendingChangesChanged);
 
-        // The override file MUST still exist — this is the load-bearing
-        // assertion: removeUserPreset cannot collateral-damage overrides.
-        QVERIFY2(QFileInfo::exists(overrideFilePath),
-                 "override file was deleted by removeUserPreset(\"zone.snapIn\") — preset CRUD MUST NOT touch override "
-                 "slots");
-        QCOMPARE(c.rawProfile(QStringLiteral("zone.snapIn")).value(QStringLiteral("duration")).toInt(), 250);
-    }
+        QVERIFY(c.setOverride(QStringLiteral("zone.snapIn"), profile));
+        QCOMPARE(overrideSpy.count(), 1);
+        QCOMPARE(pendingSpy.count(), 1);
 
-    // ─── Atomic motion-set application ────────────────────────────────────
-
-    /// Manually plant a malformed motion-set file (mixing valid and
-    /// invalid entries) and verify applyMotionSet() rejects the whole
-    /// thing rather than partially writing. Pre-fix, the loop wrote
-    /// each valid entry and skipped invalid ones, leaving inconsistent
-    /// state.
-    void applyMotionSet_malformedEntryRejectsWholeSet()
-    {
-        QTemporaryDir tmp;
-        QVERIFY(tmp.isValid());
-        AnimationsPageController c;
-        c.setUserProfilesDirOverride(tmp.path());
-
-        // Make the motion-sets directory and write a hand-crafted file
-        // with one VALID entry and one INVALID entry (unknown path).
-        const QString setsDir = tmp.path() + QStringLiteral("/motionsets");
-        QDir().mkpath(setsDir);
-        const QString setPath = setsDir + QStringLiteral("/bad-set.json");
-
-        QJsonObject validEntry;
-        validEntry.insert(QStringLiteral("path"), QStringLiteral("zone.snapIn"));
-        QJsonObject validProfile;
-        validProfile.insert(QStringLiteral("duration"), 222);
-        validEntry.insert(QStringLiteral("profile"), validProfile);
-
-        QJsonObject invalidEntry;
-        invalidEntry.insert(QStringLiteral("path"), QStringLiteral("../etc/passwd"));
-        invalidEntry.insert(QStringLiteral("profile"), QJsonObject{});
-
-        QJsonArray overrides;
-        overrides.append(validEntry);
-        overrides.append(invalidEntry);
-
-        QJsonObject root;
-        root.insert(QStringLiteral("name"), QStringLiteral("bad-set"));
-        root.insert(QStringLiteral("overrides"), overrides);
-        root.insert(QStringLiteral("version"), 1);
-
-        QFile f(setPath);
-        QVERIFY(f.open(QIODevice::WriteOnly));
-        f.write(QJsonDocument(root).toJson());
-        f.close();
-
-        // No prior override at zone.snapIn.
-        QVERIFY(!c.hasOverride(QStringLiteral("zone.snapIn")));
-
-        QVERIFY(!c.applyMotionSet(QStringLiteral("bad-set")));
-
-        // Critical: the valid entry MUST NOT have been written. Atomic
-        // semantics — all-or-nothing. Pre-fix this would be true.
-        QVERIFY2(!c.hasOverride(QStringLiteral("zone.snapIn")),
-                 "applyMotionSet wrote partial state from a malformed set — should have rejected atomically");
-    }
-
-    // ─── Shader override pendingChangesChanged emission ───────────────────
-
-    /// Pre-fix, setShaderOverride and clearShaderOverride mutated
-    /// settings but never emitted pendingChangesChanged, so the Save
-    /// button never lit up for shader edits.
-    void setShaderOverride_emitsPendingChangesChanged()
-    {
-        IsolatedConfigGuard guard;
-        Settings settings;
-        PhosphorAnimationShaders::AnimationShaderRegistry registry;
-        AnimationsPageController c(&registry, &settings);
-
-        QSignalSpy spy(&c, &AnimationsPageController::pendingChangesChanged);
-        QVERIFY(c.setShaderOverride(QStringLiteral("osd.show"), QStringLiteral("pixelate"), {}));
-        QVERIFY2(spy.count() >= 1, "setShaderOverride MUST emit pendingChangesChanged");
-    }
-
-    /// `setShaderOverride(path, "")` — the empty-effectId clear shorthand
-    /// — MUST also emit pendingChangesChanged. Pre-fix it routed to
-    /// clearShaderOverride which silently mutated.
-    void setShaderOverride_emptyEffectClearsAndEmits()
-    {
-        IsolatedConfigGuard guard;
-        Settings settings;
-        PhosphorAnimationShaders::AnimationShaderRegistry registry;
-        AnimationsPageController c(&registry, &settings);
-
-        // First write something so there's state to clear.
-        QVERIFY(c.setShaderOverride(QStringLiteral("osd.show"), QStringLiteral("pixelate"), {}));
-
-        QSignalSpy spy(&c, &AnimationsPageController::pendingChangesChanged);
-        QVERIFY(c.setShaderOverride(QStringLiteral("osd.show"), QString(), {}));
-        QVERIFY2(spy.count() >= 1, "setShaderOverride('','') MUST emit pendingChangesChanged on clear");
-        QVERIFY(c.rawShaderProfile(QStringLiteral("osd.show")).isEmpty());
-    }
-
-    /// Set then explicit clear → both calls fire pendingChangesChanged.
-    void setShaderOverride_thenClear_emitsTwice()
-    {
-        IsolatedConfigGuard guard;
-        Settings settings;
-        PhosphorAnimationShaders::AnimationShaderRegistry registry;
-        AnimationsPageController c(&registry, &settings);
-
-        QSignalSpy spy(&c, &AnimationsPageController::pendingChangesChanged);
-        QVERIFY(c.setShaderOverride(QStringLiteral("osd.show"), QStringLiteral("pixelate"), {}));
-        QVERIFY(c.clearShaderOverride(QStringLiteral("osd.show")));
-        QVERIFY2(spy.count() >= 2,
-                 qPrintable(QStringLiteral("expected ≥2 emissions, got ") + QString::number(spy.count())));
+        // Second identical write — must short-circuit, no extra signals.
+        QVERIFY(c.setOverride(QStringLiteral("zone.snapIn"), profile));
+        QCOMPARE(overrideSpy.count(), 1);
+        QCOMPARE(pendingSpy.count(), 1);
     }
 
     // ─── Shader-leg support gate ──────────────────────────────────────────
@@ -944,87 +492,6 @@ private Q_SLOTS:
         // Empty path / nonsense path.
         QVERIFY(!c.supportsShaderLeg(QString()));
         QVERIFY(!c.supportsShaderLeg(QStringLiteral("../etc/passwd")));
-    }
-
-    // ─── Pass-1 behaviour pins ────────────────────────────────────────────
-
-    /// `setOverride` compare-and-skip: writing the same Profile twice
-    /// must not fire a second `overrideChanged` / `pendingChangesChanged`.
-    /// Pre-fix the second write rewrote an identical file and lit Save.
-    void setOverride_repeatedWriteIsNoOp()
-    {
-        QTemporaryDir tmp;
-        QVERIFY(tmp.isValid());
-        AnimationsPageController c;
-        c.setUserProfilesDirOverride(tmp.path());
-
-        const auto profile = QVariantMap{{QStringLiteral("duration"), 250},
-                                         {QStringLiteral("curve"), QStringLiteral("0.33,1.00,0.68,1.00")}};
-
-        QSignalSpy overrideSpy(&c, &AnimationsPageController::overrideChanged);
-        QSignalSpy pendingSpy(&c, &AnimationsPageController::pendingChangesChanged);
-
-        QVERIFY(c.setOverride(QStringLiteral("zone.snapIn"), profile));
-        QCOMPARE(overrideSpy.count(), 1);
-        QCOMPARE(pendingSpy.count(), 1);
-
-        // Second identical write — must short-circuit, no extra signals.
-        QVERIFY(c.setOverride(QStringLiteral("zone.snapIn"), profile));
-        QCOMPARE(overrideSpy.count(), 1);
-        QCOMPARE(pendingSpy.count(), 1);
-    }
-
-    /// `setShaderOverride` short-circuits when the request matches the
-    /// existing tree state. Without this, the QML two-way binding round-
-    /// trip (combo activation → controller → settings → tree → QML
-    /// refresh → re-emit) would dirty the page on every reload.
-    void setShaderOverride_noOpWhenTreeAlreadyMatches()
-    {
-        IsolatedConfigGuard guard;
-        Settings settings;
-        PhosphorAnimationShaders::AnimationShaderRegistry registry;
-        AnimationsPageController c(&registry, &settings);
-
-        QVERIFY(c.setShaderOverride(QStringLiteral("osd.show"), QStringLiteral("pixelate"), {}));
-
-        QSignalSpy pendingSpy(&c, &AnimationsPageController::pendingChangesChanged);
-        // Same effectId + same (empty) parameters — must early-return.
-        QVERIFY(c.setShaderOverride(QStringLiteral("osd.show"), QStringLiteral("pixelate"), {}));
-        QCOMPARE(pendingSpy.count(), 0);
-    }
-
-    // ─── Logging on malformed JSON ────────────────────────────────────────
-
-    /// Plant an unparseable JSON file in the profiles dir; userPresets()
-    /// should skip it AND log a warning at the qCWarning level. The
-    /// emission is the load-bearing piece — pre-fix the parse error was
-    /// silently swallowed.
-    void userPresets_malformedJsonLogsAndSkips()
-    {
-        QTemporaryDir tmp;
-        QVERIFY(tmp.isValid());
-        AnimationsPageController c;
-        c.setUserProfilesDirOverride(tmp.path());
-
-        // Write a known-good preset so the iteration has at least one
-        // success result.
-        QVERIFY(c.addUserPreset(QStringLiteral("Good"), {{QStringLiteral("duration"), 100}}));
-
-        // Plant a malformed file directly.
-        QFile bad(tmp.path() + QStringLiteral("/garbage.json"));
-        QVERIFY(bad.open(QIODevice::WriteOnly));
-        bad.write("{ this is not valid json");
-        bad.close();
-
-        // Expect a warning to be logged for the malformed file. The
-        // exact message is "AnimationPresetLibrary: failed to parse <path> : <error>".
-        QTest::ignoreMessage(QtWarningMsg,
-                             QRegularExpression(QStringLiteral("AnimationPresetLibrary: failed to parse.*garbage")));
-
-        const QVariantList presets = c.userPresets();
-        // The good preset still surfaces; the malformed one is skipped.
-        QCOMPARE(presets.size(), 1);
-        QCOMPARE(presets.first().toMap().value(QStringLiteral("name")).toString(), QStringLiteral("Good"));
     }
 };
